Refactor Type.hpp [wip]
