+++ setup.py
@@ -27,7 +27,8 @@ setuptools.setup(
             'typed_python._types',
             ['typed_python/_runtime.cc',
              'typed_python/_types.cc',
-             'typed_python/native_instance_wrapper.cc'
+             'typed_python/native_instance_wrapper.cc',
+             'typed_python/Type.cpp'
              ],
             extra_compile_args=['-std=c++14', '-Wno-sign-compare', '-Wno-narrowing', '-Wno-unused-variable']
             )
