+++ typed_python/Type.hpp
@@ -19,6 +19,7 @@
 class Type;
 
 class None;
+
 class Bool;
 class UInt8;
 class UInt16;
@@ -28,10 +29,11 @@ class Int8;
 class Int16;
 class Int32;
 class Int64;
-class String;
-class Bytes;
 class Float32;
 class Float64;
+
+class String;
+class Bytes;
 class OneOf;
 class Value;
 class TupleOf;
@@ -119,29 +121,11 @@ public:
         return this;
     }
 
-    void repr(instance_ptr self, std::ostringstream& out) {
-        assertForwardsResolved();
+    void repr(instance_ptr self, std::ostringstream& out);
 
-        this->check([&](auto& subtype) {
-            subtype.repr(self, out);
-        });
-    }
+    char cmp(instance_ptr left, instance_ptr right);
 
-    char cmp(instance_ptr left, instance_ptr right) {
-        assertForwardsResolved();
-
-        return this->check([&](auto& subtype) {
-            return subtype.cmp(left, right);
-        });
-    }
-
-    int32_t hash32(instance_ptr left) {
-        assertForwardsResolved();
-
-        return this->check([&](auto& subtype) {
-            return subtype.hash32(left);
-        });
-    }
+    int32_t hash32(instance_ptr left);
 
     template<class buf_t>
     void serialize(instance_ptr left, buf_t& buffer) {
@@ -161,52 +145,9 @@ public:
         });
     }
 
-    void swap(instance_ptr left, instance_ptr right) {
-        assertForwardsResolved();
+    void swap(instance_ptr left, instance_ptr right);
 
-        if (left == right) {
-            return;
-        }
-
-        size_t remaining = m_size;
-        while (remaining >= 8) {
-            int64_t temp = *(int64_t*)left;
-            *(int64_t*)left = *(int64_t*)right;
-            *(int64_t*)right = temp;
-
-            remaining -= 8;
-            left += 8;
-            right += 8;
-        }
-
-        while (remaining > 0) {
-            int8_t temp = *(int8_t*)left;
-            *(int8_t*)left = *(int8_t*)right;
-            *(int8_t*)right = temp;
-
-            remaining -= 1;
-            left += 1;
-            right += 1;
-        }
-    }
-
-    static char byteCompare(uint8_t* l, uint8_t* r, size_t count) {
-        while (count >= 8 && *(uint64_t*)l == *(uint64_t*)r) {
-            l += 8;
-            r += 8;
-            count -= 8;
-        }
-
-        for (long k = 0; k < count; k++) {
-            if (l[k] < r[k]) {
-                return -1;
-            }
-            if (l[k] > r[k]) {
-                return 1;
-            }
-        }
-        return 0;
-    }
+    static char byteCompare(uint8_t* l, uint8_t* r, size_t count);
 
     template<class T>
     auto check(const T& f) -> decltype(f(*this)) {
@@ -352,17 +293,11 @@ public:
     void _visitContainedTypes(const visitor_type& v) {}
 
 
-    void constructor(instance_ptr self) {
-        assertForwardsResolved();
+    // call subtype.constructor
+    void constructor(instance_ptr self);
 
-        this->check([&](auto& subtype) { subtype.constructor(self); } );
-    }
-
-    void destroy(instance_ptr self) {
-        assertForwardsResolved();
-
-        this->check([&](auto& subtype) { subtype.destroy(self); } );
-    }
+    // call subtype.destroy
+    void destroy(instance_ptr self);
 
     template<class ptr_func>
     void destroy(int64_t count, const ptr_func& ptrToChild) {
@@ -389,35 +324,13 @@ public:
         });
     }
 
-    void forwardTypesMayHaveChanged() {
-        m_references_unresolved_forwards = false;
+    void forwardTypesMayHaveChanged();
 
-        visitReferencedTypes([&](Type* t) {
-            if (t->references_unresolved_forwards()) {
-                m_references_unresolved_forwards = true;
-            }
-        });
+    // call subtype.copy_constructor
+    void copy_constructor(instance_ptr self, instance_ptr other);
 
-        this->check([&](auto& subtype) {
-            subtype._forwardTypesMayHaveChanged();
-        });
-
-        if (mTypeRep) {
-            updateTypeRepForType(this, mTypeRep);
-        }
-    }
-
-    void copy_constructor(instance_ptr self, instance_ptr other) {
-        assertForwardsResolved();
-
-        this->check([&](auto& subtype) { subtype.copy_constructor(self, other); } );
-    }
-
-    void assign(instance_ptr self, instance_ptr other) {
-        assertForwardsResolved();
-
-        this->check([&](auto& subtype) { subtype.assign(self, other); } );
-    }
+    // call subtype.assign
+    void assign(instance_ptr self, instance_ptr other);
 
     PyTypeObject* getTypeRep() const {
         return mTypeRep;
@@ -435,35 +348,7 @@ public:
         return m_references_unresolved_forwards;
     }
 
-    bool isBinaryCompatibleWith(Type* other) {
-        if (other == this) {
-            return true;
-        }
-
-        while (other->getTypeCategory() == TypeCategory::catPythonSubclass) {
-            other = other->getBaseType();
-        }
-
-        auto it = mIsBinaryCompatible.find(other);
-        if (it != mIsBinaryCompatible.end()) {
-            return it->second != BinaryCompatibilityCategory::Incompatible;
-        }
-
-        //mark that we are recursing through this datastructure. we don't want to
-        //loop indefinitely.
-        mIsBinaryCompatible[other] = BinaryCompatibilityCategory::Checking;
-
-        bool isCompatible = this->check([&](auto& subtype) {
-            return subtype.isBinaryCompatibleWithConcrete(other);
-        });
-
-        mIsBinaryCompatible[other] = isCompatible ?
-            BinaryCompatibilityCategory::Compatible :
-            BinaryCompatibilityCategory::Incompatible
-            ;
-
-        return isCompatible;
-    }
+    bool isBinaryCompatibleWith(Type* other);
 
     bool isBinaryCompatibleWithConcrete(Type* other) {
         return false;
@@ -508,8 +393,10 @@ protected:
 
 };
 
-//forward types are never actually used - they must be removed from the graph before
-//any types that contain them can be used.
+#include "RegisterTypes.hpp"
+
+// forward types are never actually used - they must be removed from the graph before
+// any types that contain them can be used.
 class Forward : public Type {
 public:
     Forward(PyObject* deferredDefinition, std::string name) :
@@ -585,25 +472,7 @@ public:
         forwardTypesMayHaveChanged();
     }
 
-    bool isBinaryCompatibleWithConcrete(Type* other) {
-        if (other->getTypeCategory() != TypeCategory::catOneOf) {
-            return false;
-        }
-
-        OneOf* otherO = (OneOf*)other;
-
-        if (m_types.size() != otherO->m_types.size()) {
-            return false;
-        }
-
-        for (long k = 0; k < m_types.size(); k++) {
-            if (!m_types[k]->isBinaryCompatibleWith(otherO->m_types[k])) {
-                return false;
-            }
-        }
-
-        return true;
-    }
+    bool isBinaryCompatibleWithConcrete(Type* other);
 
     template<class visitor_type>
     void _visitContainedTypes(const visitor_type& visitor) {
@@ -617,37 +486,9 @@ public:
         _visitContainedTypes(visitor);
     }
 
-    void _forwardTypesMayHaveChanged() {
-        m_size = computeBytecount();
-        m_name = computeName();
+    void _forwardTypesMayHaveChanged();
 
-        m_is_default_constructible = false;
-
-        for (auto typePtr: m_types) {
-            if (typePtr->is_default_constructible()) {
-                m_is_default_constructible = true;
-                break;
-            }
-        }
-    }
-
-    std::string computeName() const {
-        std::string res = "OneOf(";
-        bool first = true;
-        for (auto t: m_types) {
-            if (first) {
-                first = false;
-            } else {
-                res += ", ";
-            }
-
-            res += t->name();
-        }
-
-        res += ")";
-
-        return res;
-    }
+    std::string computeName() const;
 
     template<class buf_t>
     void deserialize(instance_ptr self, buf_t& buffer) {
@@ -665,117 +506,31 @@ public:
         m_types[*((uint8_t*)self)]->serialize(self+1, buffer);
     }
 
-    void repr(instance_ptr self, std::ostringstream& stream) {
-        m_types[*((uint8_t*)self)]->repr(self+1, stream);
-    }
+    void repr(instance_ptr self, std::ostringstream& stream);
 
-    int32_t hash32(instance_ptr left) {
-        Hash32Accumulator acc((int)getTypeCategory());
+    int32_t hash32(instance_ptr left);
 
-        acc.add(*(uint8_t*)left);
-        acc.add(m_types[*((uint8_t*)left)]->hash32(left+1));
-
-        return acc.get();
-    }
-
-    char cmp(instance_ptr left, instance_ptr right) {
-        if (((uint8_t*)left)[0] < ((uint8_t*)right)[0]) {
-            return -1;
-        }
-        if (((uint8_t*)left)[0] > ((uint8_t*)right)[0]) {
-            return 1;
-        }
-
-        return m_types[*((uint8_t*)left)]->cmp(left+1,right+1);
-    }
+    char cmp(instance_ptr left, instance_ptr right);
 
     std::pair<Type*, instance_ptr> unwrap(instance_ptr self) {
         return std::make_pair(m_types[*(uint8_t*)self], self+1);
     }
 
-    size_t computeBytecount() const {
-        size_t res = 0;
+    size_t computeBytecount() const;
 
-        for (auto t: m_types)
-            res = std::max(res, t->bytecount());
+    void constructor(instance_ptr self);
 
-        return res + 1;
-    }
+    void destroy(instance_ptr self);
 
-    void constructor(instance_ptr self) {
-        if (!m_is_default_constructible) {
-            throw std::runtime_error(m_name + " is not default-constructible");
-        }
+    void copy_constructor(instance_ptr self, instance_ptr other);
 
-        for (size_t k = 0; k < m_types.size(); k++) {
-            if (m_types[k]->is_default_constructible()) {
-                *(uint8_t*)self = k;
-                m_types[k]->constructor(self+1);
-                return;
-            }
-        }
-    }
-
-    void destroy(instance_ptr self) {
-        uint8_t which = *(uint8_t*)(self);
-        m_types[which]->destroy(self+1);
-    }
-
-    void copy_constructor(instance_ptr self, instance_ptr other) {
-        uint8_t which = *(uint8_t*)self = *(uint8_t*)other;
-        m_types[which]->copy_constructor(self+1, other+1);
-    }
-
-    void assign(instance_ptr self, instance_ptr other) {
-        uint8_t which = *(uint8_t*)self;
-        if (which == *(uint8_t*)other) {
-            m_types[which]->assign(self+1,other+1);
-        } else {
-            m_types[which]->destroy(self+1);
-
-            uint8_t otherWhich = *(uint8_t*)other;
-            *(uint8_t*)self = otherWhich;
-            m_types[otherWhich]->copy_constructor(self+1,other+1);
-        }
-    }
+    void assign(instance_ptr self, instance_ptr other);
 
     const std::vector<Type*>& getTypes() const {
         return m_types;
     }
 
-    static OneOf* Make(const std::vector<Type*>& types) {
-        std::vector<Type*> flat_typelist;
-        std::set<Type*> seen;
-
-        //make sure we only get each type once and don't have any other 'OneOf' in there...
-        std::function<void (const std::vector<Type*>)> visit = [&](const std::vector<Type*>& subvec) {
-            for (auto t: subvec) {
-                if (t->getTypeCategory() == catOneOf) {
-                    visit( ((OneOf*)t)->getTypes() );
-                } else if (seen.find(t) == seen.end()) {
-                    flat_typelist.push_back(t);
-                    seen.insert(t);
-                }
-            }
-        };
-
-        visit(types);
-
-        static std::mutex guard;
-
-        std::lock_guard<std::mutex> lock(guard);
-
-        typedef const std::vector<Type*> keytype;
-
-        static std::map<keytype, OneOf*> m;
-
-        auto it = m.find(flat_typelist);
-        if (it == m.end()) {
-            it = m.insert(std::make_pair(flat_typelist, new OneOf(flat_typelist))).first;
-        }
-
-        return it->second;
-    }
+    static OneOf* Make(const std::vector<Type*>& types);
 
 private:
     std::vector<Type*> m_types;
@@ -794,25 +549,7 @@ public:
     {
     }
 
-    bool isBinaryCompatibleWithConcrete(Type* other) {
-        if (other->getTypeCategory() != m_typeCategory) {
-            return false;
-        }
-
-        CompositeType* otherO = (CompositeType*)other;
-
-        if (m_types.size() != otherO->m_types.size()) {
-            return false;
-        }
-
-        for (long k = 0; k < m_types.size(); k++) {
-            if (!m_types[k]->isBinaryCompatibleWith(otherO->m_types[k])) {
-                return false;
-            }
-        }
-
-        return true;
-    }
+    bool isBinaryCompatibleWithConcrete(Type* other);
 
     template<class visitor_type>
     void _visitContainedTypes(const visitor_type& visitor) {
@@ -826,37 +563,13 @@ public:
         _visitContainedTypes(visitor);
     }
 
-    void _forwardTypesMayHaveChanged() {
-        m_is_default_constructible = true;
-        m_size = 0;
-        m_byte_offsets.clear();
-
-        for (auto t: m_types) {
-            m_byte_offsets.push_back(m_size);
-            m_size += t->bytecount();
-        }
-
-        for (auto t: m_types) {
-            if (!t->is_default_constructible()) {
-                m_is_default_constructible = false;
-            }
-        }
-    }
+    void _forwardTypesMayHaveChanged();
 
     instance_ptr eltPtr(instance_ptr self, int64_t ix) const {
         return self + m_byte_offsets[ix];
     }
 
-    char cmp(instance_ptr left, instance_ptr right) {
-        for (long k = 0; k < m_types.size(); k++) {
-            char res = m_types[k]->cmp(left + m_byte_offsets[k], right + m_byte_offsets[k]);
-            if (res != 0) {
-                return res;
-            }
-        }
-
-        return 0;
-    }
+    char cmp(instance_ptr left, instance_ptr right);
 
     template<class buf_t>
     void deserialize(instance_ptr self, buf_t& buffer) {
@@ -872,38 +585,9 @@ public:
         }
     }
 
-    void repr(instance_ptr self, std::ostringstream& stream) {
-        stream << "(";
+    void repr(instance_ptr self, std::ostringstream& stream);
 
-        for (long k = 0; k < getTypes().size();k++) {
-            if (k > 0) {
-                stream << ", ";
-            }
-
-            if (k < m_names.size()) {
-                stream << m_names[k] << "=";
-            }
-
-            getTypes()[k]->repr(eltPtr(self,k),stream);
-        }
-        if (getTypes().size() == 1) {
-            stream << ",";
-        }
-
-        stream << ")";
-    }
-
-    int32_t hash32(instance_ptr left) {
-        Hash32Accumulator acc((int)getTypeCategory());
-
-        for (long k = 0; k < getTypes().size();k++) {
-            acc.add(getTypes()[k]->hash32(eltPtr(left,k)));
-        }
-
-        acc.add(getTypes().size());
-
-        return acc.get();
-    }
+    int32_t hash32(instance_ptr left);
 
     template<class sub_constructor>
     void constructor(instance_ptr self, const sub_constructor& initializer) {
@@ -919,33 +603,13 @@ public:
         }
     }
 
-    void constructor(instance_ptr self) {
-        if (!m_is_default_constructible) {
-            throw std::runtime_error(m_name + " is not default-constructible");
-        }
+    void constructor(instance_ptr self);
 
-        for (size_t k = 0; k < m_types.size(); k++) {
-            m_types[k]->constructor(self+m_byte_offsets[k]);
-        }
-    }
+    void destroy(instance_ptr self);
 
-    void destroy(instance_ptr self) {
-        for (long k = (long)m_types.size() - 1; k >= 0; k--) {
-            m_types[k]->destroy(self+m_byte_offsets[k]);
-        }
-    }
+    void copy_constructor(instance_ptr self, instance_ptr other);
 
-    void copy_constructor(instance_ptr self, instance_ptr other) {
-        for (long k = (long)m_types.size() - 1; k >= 0; k--) {
-            m_types[k]->copy_constructor(self + m_byte_offsets[k], other+m_byte_offsets[k]);
-        }
-    }
-
-    void assign(instance_ptr self, instance_ptr other) {
-        for (long k = (long)m_types.size() - 1; k >= 0; k--) {
-            m_types[k]->assign(self + m_byte_offsets[k], other+m_byte_offsets[k]);
-        }
-    }
+    void assign(instance_ptr self, instance_ptr other);
 
     const std::vector<Type*>& getTypes() const {
         return m_types;
@@ -989,20 +653,7 @@ public:
         forwardTypesMayHaveChanged();
     }
 
-    void _forwardTypesMayHaveChanged() {
-        ((CompositeType*)this)->_forwardTypesMayHaveChanged();
-
-        std::string oldName = m_name;
-
-        m_name = "NamedTuple(";
-        for (long k = 0; k < m_types.size();k++) {
-            if (k) {
-                m_name += ", ";
-            }
-            m_name += m_names[k] + "=" + m_types[k]->name();
-        }
-        m_name += ")";
-    }
+    void _forwardTypesMayHaveChanged();
 
     static NamedTuple* Make(const std::vector<Type*>& types, const std::vector<std::string>& names) {
         return MakeSubtype<NamedTuple>(types,names);
@@ -1017,18 +668,7 @@ public:
         forwardTypesMayHaveChanged();
     }
 
-    void _forwardTypesMayHaveChanged() {
-        ((CompositeType*)this)->_forwardTypesMayHaveChanged();
-
-        m_name = "Tuple(";
-        for (long k = 0; k < m_types.size();k++) {
-            if (k) {
-                m_name += ", ";
-            }
-            m_name += m_types[k]->name();
-        }
-        m_name += ")";
-    }
+    void _forwardTypesMayHaveChanged();
 
     static Tuple* Make(const std::vector<Type*>& types) {
         return MakeSubtype<Tuple>(types, std::vector<std::string>());
@@ -1055,15 +695,7 @@ public:
         forwardTypesMayHaveChanged();
     }
 
-    bool isBinaryCompatibleWithConcrete(Type* other) {
-        if (other->getTypeCategory() != m_typeCategory) {
-            return false;
-        }
-
-        TupleOf* otherO = (TupleOf*)other;
-
-        return m_element_type->isBinaryCompatibleWith(otherO->m_element_type);
-    }
+    bool isBinaryCompatibleWithConcrete(Type* other);
 
     template<class visitor_type>
     void _visitContainedTypes(const visitor_type& visitor) {
@@ -1101,127 +733,23 @@ public:
         });
     }
 
-    void repr(instance_ptr self, std::ostringstream& stream) {
-        stream << "(";
+    void repr(instance_ptr self, std::ostringstream& stream);
 
-        int32_t ct = count(self);
+    int32_t hash32(instance_ptr left);
 
-        for (long k = 0; k < ct;k++) {
-            if (k > 0) {
-                stream << ", ";
-            }
-
-            m_element_type->repr(eltPtr(self,k),stream);
-        }
-
-        stream << ")";
-    }
-
-    int32_t hash32(instance_ptr left) {
-        if (!(*(layout**)left)) {
-            return 0x123;
-        }
-
-        if ((*(layout**)left)->hash_cache == -1) {
-            Hash32Accumulator acc((int)getTypeCategory());
-
-            int32_t ct = count(left);
-            acc.add(ct);
-
-            for (long k = 0; k < ct;k++) {
-                acc.add(m_element_type->hash32(eltPtr(left, k)));
-            }
-
-            (*(layout**)left)->hash_cache = acc.get();
-            if ((*(layout**)left)->hash_cache == -1) {
-                (*(layout**)left)->hash_cache = -2;
-            }
-        }
-
-        return (*(layout**)left)->hash_cache;
-    }
-
-    char cmp(instance_ptr left, instance_ptr right) {
-        if (!(*(layout**)left) && (*(layout**)right)) {
-            return -1;
-        }
-        if (!(*(layout**)right) && (*(layout**)left)) {
-            return 1;
-        }
-        if (!(*(layout**)right) && !(*(layout**)left)) {
-            return 0;
-        }
-        layout& left_layout = **(layout**)left;
-        layout& right_layout = **(layout**)right;
-
-        if (&left_layout == &right_layout) {
-            return 0;
-        }
-
-        size_t bytesPer = m_element_type->bytecount();
-
-        for (long k = 0; k < left_layout.count && k < right_layout.count; k++) {
-            char res = m_element_type->cmp(left_layout.data + bytesPer * k,
-                                           right_layout.data + bytesPer * k);
-
-            if (res != 0) {
-                return res;
-            }
-        }
-
-        if (left_layout.count < right_layout.count) {
-            return -1;
-        }
-
-        if (left_layout.count > right_layout.count) {
-            return 1;
-        }
-
-        return 0;
-    }
+    char cmp(instance_ptr left, instance_ptr right);
 
     Type* getEltType() const {
         return m_element_type;
     }
 
-    static TupleOf* Make(Type* elt) {
-        static std::mutex guard;
+    static TupleOf* Make(Type* elt);
 
-        std::lock_guard<std::mutex> lock(guard);
+    instance_ptr eltPtr(instance_ptr self, int64_t i) const;
 
-        static std::map<Type*, TupleOf*> m;
+    int64_t count(instance_ptr self) const;
 
-        auto it = m.find(elt);
-        if (it == m.end()) {
-            it = m.insert(std::make_pair(elt, new TupleOf(elt))).first;
-        }
-
-        return it->second;
-    };
-
-    instance_ptr eltPtr(instance_ptr self, int64_t i) const {
-        if (!(*(layout**)self)) {
-            return self;
-        }
-
-        return (*(layout**)self)->data + i * m_element_type->bytecount();
-    }
-
-    int64_t count(instance_ptr self) const {
-        if (!(*(layout**)self)) {
-            return 0;
-        }
-
-        return (*(layout**)self)->count;
-    }
-
-    int64_t refcount(instance_ptr self) const {
-        if (!(*(layout**)self)) {
-            return 0;
-        }
-
-        return (*(layout**)self)->refcount;
-    }
+    int64_t refcount(instance_ptr self) const;
 
     template<class sub_constructor>
     void constructor(instance_ptr self, int64_t count, const sub_constructor& allocator) {
@@ -1249,40 +777,13 @@ public:
         }
     }
 
-    void constructor(instance_ptr self) {
-        constructor(self, 0, [](instance_ptr i, int64_t k) {});
-    }
+    void constructor(instance_ptr self);
 
-    void destroy(instance_ptr self) {
-        if (!(*(layout**)self)) {
-            return;
-        }
+    void destroy(instance_ptr self);
 
-        (*(layout**)self)->refcount--;
-        if ((*(layout**)self)->refcount == 0) {
-            m_element_type->destroy((*(layout**)self)->count, [&](int64_t k) {return eltPtr(self,k);});
-            free((*(layout**)self));
-        }
-    }
+    void copy_constructor(instance_ptr self, instance_ptr other);
 
-    void copy_constructor(instance_ptr self, instance_ptr other) {
-        (*(layout**)self) = (*(layout**)other);
-        if (*(layout**)self) {
-            (*(layout**)self)->refcount++;
-        }
-    }
-
-    void assign(instance_ptr self, instance_ptr other) {
-        layout* old = (*(layout**)self);
-
-        (*(layout**)self) = (*(layout**)other);
-
-        if (*(layout**)self) {
-            (*(layout**)self)->refcount++;
-        }
-
-        destroy((instance_ptr)&old);
-    }
+    void assign(instance_ptr self, instance_ptr other);
 
 private:
     Type* m_element_type;
@@ -1320,43 +821,11 @@ public:
         visitor(m_value);
     }
 
-    void _forwardTypesMayHaveChanged() {
-        m_name = "ConstDict(" + m_key->name() + "->" + m_value->name() + ")";
-        m_size = sizeof(void*);
-        m_is_default_constructible = true;
-        m_bytes_per_key = m_key->bytecount();
-        m_bytes_per_key_value_pair = m_key->bytecount() + m_value->bytecount();
-        m_bytes_per_key_subtree_pair = m_key->bytecount() + this->bytecount();
-        m_key_value_pair_type = Tuple::Make({m_key, m_value});
-    }
+    void _forwardTypesMayHaveChanged();
 
-    bool isBinaryCompatibleWithConcrete(Type* other) {
-        if (other->getTypeCategory() != m_typeCategory) {
-            return false;
-        }
+    bool isBinaryCompatibleWithConcrete(Type* other);
 
-        ConstDict* otherO = (ConstDict*)other;
-
-        return m_key->isBinaryCompatibleWith(otherO->m_key) &&
-            m_value->isBinaryCompatibleWith(otherO->m_value);
-    }
-
-    static ConstDict* Make(Type* key, Type* value) {
-        static std::mutex guard;
-
-        std::lock_guard<std::mutex> lock(guard);
-
-        static std::map<std::pair<Type*, Type*>, ConstDict*> m;
-
-        auto lookup_key = std::make_pair(key,value);
-
-        auto it = m.find(lookup_key);
-        if (it == m.end()) {
-            it = m.insert(std::make_pair(lookup_key, new ConstDict(key, value))).first;
-        }
-
-        return it->second;
-    };
+    static ConstDict* Make(Type* key, Type* value);
 
     template<class buf_t>
     void serialize(instance_ptr self, buf_t& buffer) {
@@ -1386,369 +855,50 @@ public:
         incKvPairCount(self, ct);
     }
 
-    void repr(instance_ptr self, std::ostringstream& stream) {
-        stream << "{";
+    void repr(instance_ptr self, std::ostringstream& stream);
 
-        int32_t ct = count(self);
-
-        for (long k = 0; k < ct;k++) {
-            if (k > 0) {
-                stream << ", ";
-            }
-
-            m_key->repr(kvPairPtrKey(self,k),stream);
-            stream << ": ";
-            m_value->repr(kvPairPtrValue(self,k),stream);
-        }
-
-        stream << "}";
-    }
-
-    int32_t hash32(instance_ptr left) {
-        if (size(left) == 0) {
-            return 0x123456;
-        }
-
-        if ((*(layout**)left)->hash_cache == -1) {
-            Hash32Accumulator acc((int)getTypeCategory());
-
-            int32_t count = size(left);
-            acc.add(count);
-            for (long k = 0; k < count;k++) {
-                acc.add(m_key->hash32(kvPairPtrKey(left,k)));
-                acc.add(m_value->hash32(kvPairPtrValue(left,k)));
-            }
-
-            (*(layout**)left)->hash_cache = acc.get();
-            if ((*(layout**)left)->hash_cache == -1) {
-                (*(layout**)left)->hash_cache = -2;
-            }
-        }
-
-        return (*(layout**)left)->hash_cache;
-    }
+    int32_t hash32(instance_ptr left);
 
     //to make this fast(er), we do dict size comparison first, then keys, then values
-    char cmp(instance_ptr left, instance_ptr right) {
-        if (size(left) < size(right)) {
-            return -1;
-        }
-        if (size(left) > size(right)) {
-            return 1;
-        }
+    char cmp(instance_ptr left, instance_ptr right);
 
-        if (*(layout**)left == *(layout**)right) {
-            return 0;
-        }
-
-        int ct = count(left);
-        for (long k = 0; k < ct; k++) {
-            char res = m_key->cmp(kvPairPtrKey(left,k), kvPairPtrKey(right,k));
-            if (res) {
-                return res;
-            }
-        }
-
-        for (long k = 0; k < ct; k++) {
-            char res = m_value->cmp(
-                kvPairPtrValue(left,k),
-                kvPairPtrValue(right,k)
-                );
-
-            if (res) {
-                return res;
-            }
-        }
-
-        return 0;
-    }
-
-    void addDicts(instance_ptr lhs, instance_ptr rhs, instance_ptr output) const {
-        std::vector<instance_ptr> keep;
-
-        int64_t lhsCount = count(lhs);
-        int64_t rhsCount = count(rhs);
-
-        for (long k = 0; k < lhsCount; k++) {
-            instance_ptr lhsVal = kvPairPtrKey(lhs, k);
-
-            if (!lookupValueByKey(rhs, lhsVal)) {
-                keep.push_back(lhsVal);
-            }
-        }
-
-        constructor(output, rhsCount + keep.size(), false);
-
-        for (long k = 0; k < rhsCount; k++) {
-            m_key->copy_constructor(kvPairPtrKey(output,k), kvPairPtrKey(rhs, k));
-            m_value->copy_constructor(kvPairPtrValue(output,k), kvPairPtrValue(rhs, k));
-        }
-        for (long k = 0; k < keep.size(); k++) {
-            m_key->copy_constructor(kvPairPtrKey(output,k + rhsCount), keep[k]);
-            m_value->copy_constructor(kvPairPtrValue(output,k + rhsCount), keep[k] + m_bytes_per_key);
-        }
-        incKvPairCount(output, keep.size() + rhsCount);
-
-        sortKvPairs(output);
-    }
+    void addDicts(instance_ptr lhs, instance_ptr rhs, instance_ptr output) const;
 
     TupleOf* tupleOfKeysType() const {
         return TupleOf::Make(m_key);
     }
 
-    void subtractTupleOfKeysFromDict(instance_ptr lhs, instance_ptr rhs, instance_ptr output) const {
-        TupleOf* tupleType = tupleOfKeysType();
+    void subtractTupleOfKeysFromDict(instance_ptr lhs, instance_ptr rhs, instance_ptr output) const;
 
-        int64_t lhsCount = count(lhs);
-        int64_t rhsCount = tupleType->count(rhs);
+    instance_ptr kvPairPtrKey(instance_ptr self, int64_t i) const;
 
-        std::set<int> remove;
+    instance_ptr kvPairPtrValue(instance_ptr self, int64_t i) const;
 
-        for (long k = 0; k < rhsCount; k++) {
-            int64_t index = lookupIndexByKey(lhs, tupleType->eltPtr(rhs, k));
-            if (index != -1) {
-                remove.insert(index);
-            }
-        }
+    void incKvPairCount(instance_ptr self, int by = 1) const;
 
-        constructor(output, lhsCount - remove.size(), false);
+    void sortKvPairs(instance_ptr self) const;
 
-        long written = 0;
-        for (long k = 0; k < lhsCount; k++) {
-            if (remove.find(k) == remove.end()) {
-                m_key->copy_constructor(kvPairPtrKey(output,written), kvPairPtrKey(lhs, k));
-                m_value->copy_constructor(kvPairPtrValue(output,written), kvPairPtrValue(lhs, k));
+    instance_ptr keyTreePtr(instance_ptr self, int64_t i) const;
 
-                written++;
-            }
-        }
+    bool instanceIsSubtrees(instance_ptr self) const;
 
-        incKvPairCount(output, written);
-    }
+    int64_t count(instance_ptr self) const;
 
-    instance_ptr kvPairPtrKey(instance_ptr self, int64_t i) const {
-        if (!(*(layout**)self)) {
-            return self;
-        }
+    int64_t size(instance_ptr self) const;
 
-        layout& record = **(layout**)self;
+    int64_t lookupIndexByKey(instance_ptr self, instance_ptr key) const;
 
-        return record.data + m_bytes_per_key_value_pair * i;
-    }
+    instance_ptr lookupValueByKey(instance_ptr self, instance_ptr key) const;
 
-    instance_ptr kvPairPtrValue(instance_ptr self, int64_t i) const {
-        if (!(*(layout**)self)) {
-            return self;
-        }
+    void constructor(instance_ptr self, int64_t space, bool isPointerTree) const;
 
-        layout& record = **(layout**)self;
+    void constructor(instance_ptr self);
 
-        return record.data + m_bytes_per_key_value_pair * i + m_bytes_per_key;
-    }
+    void destroy(instance_ptr self);
 
-    void incKvPairCount(instance_ptr self, int by = 1) const {
-        if (by == 0) {
-            return;
-        }
+    void copy_constructor(instance_ptr self, instance_ptr other);
 
-        layout& record = **(layout**)self;
-        record.count += by;
-    }
-
-    void sortKvPairs(instance_ptr self) const {
-        if (!*(layout**)self) {
-            return;
-        }
-
-        layout& record = **(layout**)self;
-
-        assert(!record.subpointers);
-
-        if (record.count <= 1) {
-            return;
-        }
-        else if (record.count == 2) {
-            if (m_key->cmp(kvPairPtrKey(self, 0), kvPairPtrKey(self,1)) > 0) {
-                m_key->swap(kvPairPtrKey(self,0), kvPairPtrKey(self,1));
-                m_value->swap(kvPairPtrValue(self,0), kvPairPtrValue(self,1));
-            }
-            return;
-        } else {
-            std::vector<int> indices;
-            for (long k=0;k<record.count;k++) {
-                indices.push_back(k);
-            }
-
-            std::sort(indices.begin(), indices.end(), [&](int l, int r) {
-                char res = m_key->cmp(kvPairPtrKey(self,l),kvPairPtrKey(self,r));
-                return res < 0;
-                });
-
-            //create a temporary buffer
-            std::vector<uint8_t> d;
-            d.resize(m_bytes_per_key_value_pair * record.count);
-
-            //final_lookup contains the location of each value in the original sort
-            for (long k = 0; k < indices.size(); k++) {
-                m_key->swap(kvPairPtrKey(self, indices[k]), &d[m_bytes_per_key_value_pair*k]);
-                m_value->swap(kvPairPtrValue(self, indices[k]), &d[m_bytes_per_key_value_pair*k+m_bytes_per_key]);
-            }
-
-            //now move them back
-            for (long k = 0; k < indices.size(); k++) {
-                m_key->swap(kvPairPtrKey(self, k), &d[m_bytes_per_key_value_pair*k]);
-                m_value->swap(kvPairPtrValue(self, k), &d[m_bytes_per_key_value_pair*k+m_bytes_per_key]);
-            }
-        }
-    }
-
-    instance_ptr keyTreePtr(instance_ptr self, int64_t i) const {
-        if (!(*(layout**)self)) {
-            return self;
-        }
-
-        layout& record = **(layout**)self;
-
-        return record.data + m_bytes_per_key_subtree_pair * i;
-    }
-
-    bool instanceIsSubtrees(instance_ptr self) const {
-        if (!(*(layout**)self)) {
-            return self;
-        }
-
-        layout& record = **(layout**)self;
-
-        return record.subpointers != 0;
-    }
-
-    int64_t count(instance_ptr self) const {
-        if (!(*(layout**)self)) {
-            return 0;
-        }
-
-        layout& record = **(layout**)self;
-
-        if (record.subpointers) {
-            return record.subpointers;
-        }
-
-        return record.count;
-    }
-
-    int64_t size(instance_ptr self) const {
-        if (!(*(layout**)self)) {
-            return 0;
-        }
-
-        return (*(layout**)self)->count;
-    }
-
-    int64_t lookupIndexByKey(instance_ptr self, instance_ptr key) const {
-        if (!(*(layout**)self)) {
-            return -1;
-        }
-
-        layout& record = **(layout**)self;
-
-        assert(record.subpointers == 0); //this is not implemented yet
-
-        long low = 0;
-        long high = record.count;
-
-        while (low < high) {
-            long mid = (low+high)/2;
-            char res = m_key->cmp(kvPairPtrKey(self, mid), key);
-
-            if (res == 0) {
-                return mid;
-            } else if (res < 0) {
-                low = mid+1;
-            } else {
-                high = mid;
-            }
-        }
-
-        return -1;
-    }
-
-    instance_ptr lookupValueByKey(instance_ptr self, instance_ptr key) const {
-        int64_t offset = lookupIndexByKey(self, key);
-        if (offset == -1) {
-            return 0;
-        }
-        return kvPairPtrValue(self, offset);
-    }
-
-    void constructor(instance_ptr self, int64_t space, bool isPointerTree) const {
-        if (space == 0) {
-            (*(layout**)self) = nullptr;
-            return;
-        }
-
-        int bytesPer = isPointerTree ? m_bytes_per_key_subtree_pair : m_bytes_per_key_value_pair;
-
-        (*(layout**)self) = (layout*)malloc(sizeof(layout) + bytesPer * space);
-
-        layout& record = **(layout**)self;
-
-        record.count = 0;
-        record.subpointers = 0;
-        record.refcount = 1;
-        record.hash_cache = -1;
-    }
-
-    void constructor(instance_ptr self) {
-        (*(layout**)self) = nullptr;
-    }
-
-    void destroy(instance_ptr self) {
-        if (!(*(layout**)self)) {
-            return;
-        }
-
-        layout& record = **(layout**)self;
-
-        record.refcount--;
-        if (record.refcount == 0) {
-            if (record.subpointers == 0) {
-                m_key->destroy(record.count, [&](long ix) {
-                    return record.data + m_bytes_per_key_value_pair * ix;
-                });
-                m_value->destroy(record.count, [&](long ix) {
-                    return record.data + m_bytes_per_key_value_pair * ix + m_bytes_per_key;
-                });
-            } else {
-                m_key->destroy(record.subpointers, [&](long ix) {
-                    return record.data + m_bytes_per_key_subtree_pair * ix;
-                });
-                ((Type*)this)->destroy(record.subpointers, [&](long ix) {
-                    return record.data + m_bytes_per_key_subtree_pair * ix + m_bytes_per_key;
-                });
-            }
-
-            free((*(layout**)self));
-        }
-    }
-
-    void copy_constructor(instance_ptr self, instance_ptr other) {
-        (*(layout**)self) = (*(layout**)other);
-        if (*(layout**)self) {
-            (*(layout**)self)->refcount++;
-        }
-    }
-
-    void assign(instance_ptr self, instance_ptr other) {
-        layout* old = (*(layout**)self);
-
-        (*(layout**)self) = (*(layout**)other);
-
-        if (*(layout**)self) {
-            (*(layout**)self)->refcount++;
-        }
-
-        destroy((instance_ptr)&old);
-    }
+    void assign(instance_ptr self, instance_ptr other);
 
 
     Type* keyValuePairType() const { return m_key_value_pair_type; }
@@ -1821,229 +971,7 @@ public:
     }
 };
 
-template<class T>
-class RegisterType : public Type {
-public:
-    RegisterType(TypeCategory kind) : Type(kind)
-    {
-        m_size = sizeof(T);
-        m_is_default_constructible = true;
-    }
-
-    bool isBinaryCompatibleWithConcrete(Type* other) {
-        if (other->getTypeCategory() != m_typeCategory) {
-            return false;
-        }
-
-        return true;
-    }
-
-    void _forwardTypesMayHaveChanged() {}
-
-    template<class visitor_type>
-    void _visitReferencedTypes(const visitor_type& v) {}
-
-    template<class visitor_type>
-    void _visitContainedTypes(const visitor_type& v) {}
-
-    char cmp(instance_ptr left, instance_ptr right) {
-        if ( (*(T*)left) < (*(T*)right) ) {
-            return -1;
-        }
-        if ( (*(T*)left) > (*(T*)right) ) {
-            return 1;
-        }
-
-        return 0;
-    }
-
-    int32_t hash32(instance_ptr left) {
-        Hash32Accumulator acc((int)getTypeCategory());
-
-        acc.addRegister(*(T*)left);
-
-        return acc.get();
-    }
-
-    void constructor(instance_ptr self) {
-        new ((T*)self) T();
-    }
-
-    void destroy(instance_ptr self) {}
-
-    void copy_constructor(instance_ptr self, instance_ptr other) {
-        *((T*)self) = *((T*)other);
-    }
-
-    void assign(instance_ptr self, instance_ptr other) {
-        *((T*)self) = *((T*)other);
-    }
-
-    template<class buf_t>
-    void deserialize(instance_ptr self, buf_t& buffer) {
-        buffer.read_bytes(self, m_size);
-    }
-
-    template<class buf_t>
-    void serialize(instance_ptr self, buf_t& buffer) {
-        buffer.write_bytes(self, m_size);
-    }
-};
-
-class Bool : public RegisterType<bool> {
-public:
-    Bool() : RegisterType(TypeCategory::catBool)
-    {
-        m_name = "Bool";
-    }
-
-    void repr(instance_ptr self, std::ostringstream& stream) {
-        stream << (*(bool*)self ? "True":"False");
-    }
-
-    static Bool* Make() { static Bool res; return &res; }
-};
-
-class UInt8 : public RegisterType<uint8_t> {
-public:
-    UInt8() : RegisterType(TypeCategory::catUInt8)
-    {
-        m_name = "UInt8";
-    }
-
-    void repr(instance_ptr self, std::ostringstream& stream) {
-        stream << (uint64_t)*(uint8_t*)self << "u8";
-    }
-
-    static UInt8* Make() { static UInt8 res; return &res; }
-};
-
-class UInt16 : public RegisterType<uint16_t> {
-public:
-    UInt16() : RegisterType(TypeCategory::catUInt16)
-    {
-        m_name = "UInt16";
-    }
-
-    void repr(instance_ptr self, std::ostringstream& stream) {
-        stream << (uint64_t)*(uint16_t*)self << "u16";
-    }
-
-    static UInt16* Make() { static UInt16 res; return &res; }
-};
-
-class UInt32 : public RegisterType<uint32_t> {
-public:
-    UInt32() : RegisterType(TypeCategory::catUInt32)
-    {
-        m_name = "UInt32";
-    }
-
-    void repr(instance_ptr self, std::ostringstream& stream) {
-        stream << (uint64_t)*(uint32_t*)self << "u32";
-    }
-
-    static UInt32* Make() { static UInt32 res; return &res; }
-};
-
-class UInt64 : public RegisterType<uint64_t> {
-public:
-    UInt64() : RegisterType(TypeCategory::catUInt64)
-    {
-        m_name = "UInt64";
-    }
-
-    void repr(instance_ptr self, std::ostringstream& stream) {
-        stream << *(uint64_t*)self << "u64";
-    }
-
-    static UInt64* Make() { static UInt64 res; return &res; }
-};
-
-class Int8 : public RegisterType<int8_t> {
-public:
-    Int8() : RegisterType(TypeCategory::catInt8)
-    {
-        m_name = "Int8";
-        m_size = 1;
-    }
-
-    void repr(instance_ptr self, std::ostringstream& stream) {
-        stream << (int64_t)*(int8_t*)self << "i8";
-    }
-
-    static Int8* Make() { static Int8 res; return &res; }
-};
-
-class Int16 : public RegisterType<int16_t> {
-public:
-    Int16() : RegisterType(TypeCategory::catInt16)
-    {
-        m_name = "Int16";
-    }
-
-    void repr(instance_ptr self, std::ostringstream& stream) {
-        stream << (int64_t)*(int16_t*)self << "i16";
-    }
-
-    static Int16* Make() { static Int16 res; return &res; }
-};
-
-class Int32 : public RegisterType<int32_t> {
-public:
-    Int32() : RegisterType(TypeCategory::catInt32)
-    {
-        m_name = "Int32";
-    }
-
-    void repr(instance_ptr self, std::ostringstream& stream) {
-        stream << (int64_t)*(int32_t*)self << "i32";
-    }
-
-    static Int32* Make() { static Int32 res; return &res; }
-};
-
-class Int64 : public RegisterType<int64_t> {
-public:
-    Int64() : RegisterType(TypeCategory::catInt64)
-    {
-        m_name = "Int64";
-    }
-
-    void repr(instance_ptr self, std::ostringstream& stream) {
-        stream << *(int64_t*)self;
-    }
-
-    static Int64* Make() { static Int64 res; return &res; }
-};
-
-class Float32 : public RegisterType<float> {
-public:
-    Float32() : RegisterType(TypeCategory::catFloat32)
-    {
-        m_name = "Float32";
-    }
-
-    void repr(instance_ptr self, std::ostringstream& stream) {
-        stream << *(float*)self << "f32";
-    }
-
-    static Float32* Make() { static Float32 res; return &res; }
-};
-
-class Float64 : public RegisterType<double> {
-public:
-    Float64() : RegisterType(TypeCategory::catFloat64)
-    {
-        m_name = "Float64";
-    }
-
-    void repr(instance_ptr self, std::ostringstream& stream) {
-        stream << *(double*)self;
-    }
-
-    static Float64* Make() { static Float64 res; return &res; }
-};
+#include "Instance.hpp"
 
 class String : public Type {
 public:
@@ -2114,166 +1042,29 @@ public:
         }
     }
 
-    int32_t hash32(instance_ptr left) {
-        if (!(*(layout**)left)) {
-            return 0x12345;
-        }
+    int32_t hash32(instance_ptr left);
 
-        if ((*(layout**)left)->hash_cache == -1) {
-            Hash32Accumulator acc((int)getTypeCategory());
-            acc.addBytes(eltPtr(left, 0), bytes_per_codepoint(left) * count(left));
-            (*(layout**)left)->hash_cache = acc.get();
-            if ((*(layout**)left)->hash_cache == -1) {
-                (*(layout**)left)->hash_cache = -2;
-            }
-        }
+    char cmp(instance_ptr left, instance_ptr right);
 
-        return (*(layout**)left)->hash_cache;
-    }
+    void constructor(instance_ptr self, int64_t bytes_per_codepoint, int64_t count, const char* data) const;
 
-    char cmp(instance_ptr left, instance_ptr right) {
-        if ( !(*(layout**)left) && !(*(layout**)right) ) {
-            return 0;
-        }
-        if ( !(*(layout**)left) && (*(layout**)right) ) {
-            return -1;
-        }
-        if ( (*(layout**)left) && !(*(layout**)right) ) {
-            return 1;
-        }
+    void repr(instance_ptr self, std::ostringstream& stream);
 
-        if (bytes_per_codepoint(left) < bytes_per_codepoint(right)) {
-            return -1;
-        }
+    instance_ptr eltPtr(instance_ptr self, int64_t i) const;
 
-        if (bytes_per_codepoint(left) > bytes_per_codepoint(right)) {
-            return 1;
-        }
+    int64_t bytes_per_codepoint(instance_ptr self) const;
 
-        int bytesPer = bytes_per_codepoint(right);
-
-        char res = byteCompare(
-            eltPtr(left, 0),
-            eltPtr(right, 0),
-            bytesPer * std::min(count(left), count(right))
-            );
-
-        if (res) {
-            return res;
-        }
-
-        if (count(left) < count(right)) {
-            return -1;
-        }
-
-        if (count(left) > count(right)) {
-            return 1;
-        }
-
-        return 0;
-    }
-
-    void constructor(instance_ptr self, int64_t bytes_per_codepoint, int64_t count, const char* data) const {
-        if (count == 0) {
-            *(layout**)self = nullptr;
-            return;
-        }
-
-        (*(layout**)self) = (layout*)malloc(sizeof(layout) + count * bytes_per_codepoint);
-
-        (*(layout**)self)->bytes_per_codepoint = bytes_per_codepoint;
-        (*(layout**)self)->pointcount = count;
-        (*(layout**)self)->hash_cache = -1;
-        (*(layout**)self)->refcount = 1;
-
-        if (data) {
-            ::memcpy((*(layout**)self)->data, data, count * bytes_per_codepoint);
-        }
-    }
-
-    void repr(instance_ptr self, std::ostringstream& stream) {
-        //as if it were bytes, which is totally wrong...
-        stream << "\"";
-        long bytes = count(self);
-        uint8_t* base = eltPtr(self,0);
-
-        static char hexDigits[] = "0123456789abcdef";
-
-        for (long k = 0; k < bytes;k++) {
-            if (base[k] == '"') {
-                stream << "\\\"";
-            } else if (base[k] == '\\') {
-                stream << "\\\\";
-            } else if (isprint(base[k])) {
-                stream << base[k];
-            } else {
-                stream << "\\x" << hexDigits[base[k] / 16] << hexDigits[base[k] % 16];
-            }
-        }
-
-        stream << "\"";
-    }
-
-    instance_ptr eltPtr(instance_ptr self, int64_t i) const {
-        const static char* emptyPtr = "";
-
-        if (*(layout**)self == nullptr) {
-            return (instance_ptr)emptyPtr;
-        }
-
-        return (*(layout**)self)->eltPtr(i);
-    }
-
-    int64_t bytes_per_codepoint(instance_ptr self) const {
-        if (*(layout**)self == nullptr) {
-            return 1;
-        }
-
-        return (*(layout**)self)->bytes_per_codepoint;
-    }
-
-    int64_t count(instance_ptr self) const {
-        if (*(layout**)self == nullptr) {
-            return 0;
-        }
-
-        return (*(layout**)self)->pointcount;
-    }
+    int64_t count(instance_ptr self) const;
 
     void constructor(instance_ptr self) {
         *(layout**)self = 0;
     }
 
-    void destroy(instance_ptr self) {
-        if (!*(layout**)self) {
-            return;
-        }
+    void destroy(instance_ptr self);
 
-        (*(layout**)self)->refcount--;
+    void copy_constructor(instance_ptr self, instance_ptr other);
 
-        if ((*(layout**)self)->refcount == 0) {
-            free((*(layout**)self));
-        }
-    }
-
-    void copy_constructor(instance_ptr self, instance_ptr other) {
-        (*(layout**)self) = (*(layout**)other);
-        if (*(layout**)self) {
-            (*(layout**)self)->refcount++;
-        }
-    }
-
-    void assign(instance_ptr self, instance_ptr other) {
-        layout* old = (*(layout**)self);
-
-        (*(layout**)self) = (*(layout**)other);
-
-        if (*(layout**)self) {
-            (*(layout**)self)->refcount++;
-        }
-
-        destroy((instance_ptr)&old);
-    }
+    void assign(instance_ptr self, instance_ptr other);
 };
 
 class Bytes : public Type {
@@ -2286,41 +1077,16 @@ public:
         uint8_t data[];
     };
 
-    bool isBinaryCompatibleWithConcrete(Type* other) {
-        if (other->getTypeCategory() != m_typeCategory) {
-            return false;
-        }
-
-        return true;
+    Bytes() : Type(TypeCategory::catBytes)
+    {
+        m_name = "Bytes";
+        m_is_default_constructible = true;
+        m_size = sizeof(layout*);
     }
 
-    void repr(instance_ptr self, std::ostringstream& stream) {
-        stream << "b" << "'";
-        long bytes = count(self);
-        uint8_t* base = eltPtr(self,0);
+    bool isBinaryCompatibleWithConcrete(Type* other);
 
-        static char hexDigits[] = "0123456789abcdef";
-
-        for (long k = 0; k < bytes;k++) {
-            if (base[k] == '\'') {
-                stream << "\\'";
-            } else if (base[k] == '\r') {
-                stream << "\\r";
-            } else if (base[k] == '\n') {
-                stream << "\\n";
-            } else if (base[k] == '\t') {
-                stream << "\\t";
-            } else if (base[k] == '\\') {
-                stream << "\\\\";
-            } else if (isprint(base[k])) {
-                stream << base[k];
-            } else {
-                stream << "\\x" << hexDigits[base[k] / 16] << hexDigits[base[k] % 16];
-            }
-        }
-
-        stream << "'";
-    }
+    void repr(instance_ptr self, std::ostringstream& stream);
 
     template<class buf_t>
     void serialize(instance_ptr self, buf_t& buffer) {
@@ -2351,290 +1117,27 @@ public:
     template<class visitor_type>
     void _visitContainedTypes(const visitor_type& v) {}
 
-    int32_t hash32(instance_ptr left) {
-        Hash32Accumulator acc((int)getTypeCategory());
+    int32_t hash32(instance_ptr left);
 
-        if (!(*(layout**)left)) {
-            return 0x1234;
-        }
-
-        if ((*(layout**)left)->hash_cache == -1) {
-            Hash32Accumulator acc((int)getTypeCategory());
-
-            acc.addBytes(eltPtr(left, 0), count(left));
-
-            (*(layout**)left)->hash_cache = acc.get();
-            if ((*(layout**)left)->hash_cache == -1) {
-                (*(layout**)left)->hash_cache = -2;
-            }
-        }
-
-        return (*(layout**)left)->hash_cache;
-    }
-
-    char cmp(instance_ptr left, instance_ptr right) {
-        if ( !(*(layout**)left) && !(*(layout**)right) ) {
-            return 0;
-        }
-        if ( !(*(layout**)left) && (*(layout**)right) ) {
-            return -1;
-        }
-        if ( (*(layout**)left) && !(*(layout**)right) ) {
-            return 1;
-        }
-        if ( (*(layout**)left) == (*(layout**)right) ) {
-            return 0;
-        }
-
-        char res = byteCompare(eltPtr(left, 0), eltPtr(right, 0), std::min(count(left), count(right)));
-
-        if (res) {
-            return res;
-        }
-
-        if (count(left) < count(right)) {
-            return -1;
-        }
-
-        if (count(left) > count(right)) {
-            return 1;
-        }
-
-        return 0;
-    }
-
-    Bytes() : Type(TypeCategory::catBytes)
-    {
-        m_name = "Bytes";
-        m_is_default_constructible = true;
-        m_size = sizeof(layout*);
-    }
+    char cmp(instance_ptr left, instance_ptr right);
 
     static Bytes* Make() { static Bytes res; return &res; }
 
-    void constructor(instance_ptr self, int64_t count, const char* data) const {
-        if (count == 0) {
-            *(layout**)self = nullptr;
-            return;
-        }
-        (*(layout**)self) = (layout*)malloc(sizeof(layout) + count);
+    void constructor(instance_ptr self, int64_t count, const char* data) const;
 
-        (*(layout**)self)->bytecount = count;
-        (*(layout**)self)->refcount = 1;
-        (*(layout**)self)->hash_cache = -1;
+    instance_ptr eltPtr(instance_ptr self, int64_t i) const;
 
-        if (data) {
-            ::memcpy((*(layout**)self)->data, data, count);
-        }
-    }
-
-    instance_ptr eltPtr(instance_ptr self, int64_t i) const {
-        //we don't want to have to return null here, but there is no actual memory to back this.
-        if (*(layout**)self == nullptr) {
-            return self;
-        }
-
-        return (*(layout**)self)->data + i;
-    }
-
-    int64_t count(instance_ptr self) const {
-        if (*(layout**)self == nullptr) {
-            return 0;
-        }
-
-        return (*(layout**)self)->bytecount;
-    }
+    int64_t count(instance_ptr self) const;
 
     void constructor(instance_ptr self) {
         *(layout**)self = 0;
     }
 
-    void destroy(instance_ptr self) {
-        if (!*(layout**)self) {
-            return;
-        }
+    void destroy(instance_ptr self);
 
-        (*(layout**)self)->refcount--;
+    void copy_constructor(instance_ptr self, instance_ptr other);
 
-        if ((*(layout**)self)->refcount == 0) {
-            free((*(layout**)self));
-        }
-    }
-
-    void copy_constructor(instance_ptr self, instance_ptr other) {
-        (*(layout**)self) = (*(layout**)other);
-        if (*(layout**)self) {
-            (*(layout**)self)->refcount++;
-        }
-    }
-
-    void assign(instance_ptr self, instance_ptr other) {
-        layout* old = (*(layout**)self);
-
-        (*(layout**)self) = (*(layout**)other);
-
-        if (*(layout**)self) {
-            (*(layout**)self)->refcount++;
-        }
-
-        destroy((instance_ptr)&old);
-    }
-};
-
-class Instance {
-private:
-    class layout {
-    public:
-        std::atomic<int64_t> refcount;
-        Type* type;
-        uint8_t data[];
-    };
-
-    Instance(layout* l) :
-        mLayout(l)
-    {
-    }
-
-    static layout* allocateNoneLayout() {
-        layout* result = (layout*)malloc(sizeof(layout));
-        result->refcount = 0;
-        result->type = None::Make();
-
-        return result;
-    }
-
-    static layout* noneLayout() {
-        static layout* noneLayout = allocateNoneLayout();
-        noneLayout->refcount++;
-
-        return noneLayout;
-    }
-
-public:
-    static Instance deserialized(Type* t, DeserializationBuffer& buf) {
-        t->assertForwardsResolved();
-
-        return createAndInitialize(t, [&](instance_ptr tgt) {
-            t->deserialize(tgt, buf);
-        });
-    }
-
-    static Instance create(Type*t, instance_ptr data) {
-        t->assertForwardsResolved();
-
-        return createAndInitialize(t, [&](instance_ptr tgt) {
-            t->copy_constructor(tgt, data);
-        });
-    }
-
-    template<class initializer_type>
-    static Instance createAndInitialize(Type* t, const initializer_type& initFun) {
-        return Instance(t, initFun);
-    }
-
-    Instance() {
-        //by default, None
-        mLayout = noneLayout();
-        mLayout->refcount++;
-    }
-
-    Instance(const Instance& other) : mLayout(other.mLayout) {
-        mLayout->refcount++;
-    }
-
-    Instance(instance_ptr p, Type* t) : mLayout(nullptr) {
-        t->assertForwardsResolved();
-
-        layout* l = (layout*)malloc(sizeof(layout) + t->bytecount());
-
-        try {
-            t->copy_constructor(l->data, p);
-        } catch(...) {
-            free(l);
-            throw;
-        }
-
-        l->refcount = 1;
-        l->type = t;
-
-        mLayout = l;
-    }
-
-    template<class initializer_type>
-    Instance(Type* t, const initializer_type& initFun) : mLayout(nullptr) {
-        t->assertForwardsResolved();
-
-        layout* l = (layout*)malloc(sizeof(layout) + t->bytecount());
-
-        try {
-            initFun(l->data);
-        } catch(...) {
-            free(l);
-            throw;
-        }
-
-        l->refcount = 1;
-        l->type = t;
-
-        mLayout = l;
-    }
-
-    ~Instance() {
-        mLayout->refcount--;
-        if (mLayout->refcount == 0) {
-            mLayout->type->destroy(mLayout->data);
-            free(mLayout);
-        }
-    }
-
-    Instance& operator=(const Instance& other) {
-        other.mLayout->refcount++;
-
-        mLayout->refcount--;
-        if (mLayout->refcount == 0) {
-            mLayout->type->destroy(mLayout->data);
-            free(mLayout);
-        }
-
-        mLayout = other.mLayout;
-        return *this;
-    }
-
-    bool operator<(const Instance& other) const {
-        if (mLayout->type < other.mLayout->type) {
-            return true;
-        }
-        if (mLayout->type > other.mLayout->type) {
-            return false;
-        }
-        return mLayout->type->cmp(mLayout->data, other.mLayout->data) < 0;
-    }
-
-    std::string repr() const {
-        std::ostringstream s;
-        s << std::showpoint;
-        mLayout->type->repr(mLayout->data, s);
-        return s.str();
-    }
-
-    int32_t hash32() const {
-        return mLayout->type->hash32(mLayout->data);
-    }
-
-    Type* type() const {
-        return mLayout->type;
-    }
-
-    instance_ptr data() const {
-        return mLayout->data;
-    }
-
-    int64_t refcount() const {
-        return mLayout->refcount;
-    }
-
-private:
-    layout* mLayout;
+    void assign(instance_ptr self, instance_ptr other);
 };
 
 class Value : public Type {
@@ -2766,32 +1269,7 @@ public:
         forwardTypesMayHaveChanged();
     }
 
-    bool isBinaryCompatibleWithConcrete(Type* other) {
-        if (other->getTypeCategory() == TypeCategory::catConcreteAlternative) {
-            other = other->getBaseType();
-        }
-
-        if (other->getTypeCategory() != m_typeCategory) {
-            return false;
-        }
-
-        Alternative* otherO = (Alternative*)other;
-
-        if (m_subtypes.size() != otherO->m_subtypes.size()) {
-            return false;
-        }
-
-        for (long k = 0; k < m_subtypes.size(); k++) {
-            if (m_subtypes[k].first != otherO->m_subtypes[k].first) {
-                return false;
-            }
-            if (!m_subtypes[k].second->isBinaryCompatibleWith(otherO->m_subtypes[k].second)) {
-                return false;
-            }
-        }
-
-        return true;
-    }
+    bool isBinaryCompatibleWithConcrete(Type* other);
 
     template<class visitor_type>
     void _visitContainedTypes(const visitor_type& visitor) {
@@ -2811,63 +1289,9 @@ public:
         }
     }
 
-    void _forwardTypesMayHaveChanged() {
-        m_size = sizeof(void*);
+    void _forwardTypesMayHaveChanged();
 
-        m_is_default_constructible = false;
-        m_all_alternatives_empty = true;
-        m_arg_positions.clear();
-        m_default_construction_ix = 0;
-        m_default_construction_type = nullptr;
-
-        for (auto& subtype_pair: m_subtypes) {
-            if (subtype_pair.second->bytecount() > 0) {
-                m_all_alternatives_empty = false;
-            }
-
-            if (m_arg_positions.find(subtype_pair.first) != m_arg_positions.end()) {
-                throw std::runtime_error("Can't create an alternative with " +
-                        subtype_pair.first + " defined twice.");
-            }
-
-            m_arg_positions[subtype_pair.first] = m_arg_positions.size();
-
-            if (subtype_pair.second->is_default_constructible() && !m_is_default_constructible) {
-                m_is_default_constructible = true;
-                m_default_construction_ix = m_arg_positions[subtype_pair.first];
-            }
-        }
-
-        m_size = (m_all_alternatives_empty ? 1 : sizeof(void*));
-    }
-
-    char cmp(instance_ptr left, instance_ptr right) {
-        if (m_all_alternatives_empty) {
-            if (*(uint8_t*)left < *(uint8_t*)right) {
-                return -1;
-            }
-            if (*(uint8_t*)left > *(uint8_t*)right) {
-                return 1;
-            }
-            return 0;
-        }
-
-        layout& record_l = **(layout**)left;
-        layout& record_r = **(layout**)right;
-
-        if ( &record_l == &record_r ) {
-            return 0;
-        }
-
-        if (record_l.which < record_r.which) {
-            return -1;
-        }
-        if (record_l.which > record_r.which) {
-            return 1;
-        }
-
-        return m_subtypes[record_l.which].second->cmp(record_l.data, record_r.data);
-    }
+    char cmp(instance_ptr left, instance_ptr right);
 
     template<class buf_t>
     void serialize(instance_ptr self, buf_t& buffer) {
@@ -2900,106 +1324,26 @@ public:
         m_subtypes[w].second->deserialize(record.data, buffer);
     }
 
-    void repr(instance_ptr self, std::ostringstream& stream) {
-        stream << m_subtypes[which(self)].first;
-        m_subtypes[which(self)].second->repr(eltPtr(self), stream);
-    }
+    void repr(instance_ptr self, std::ostringstream& stream);
 
-    int32_t hash32(instance_ptr left) {
-        Hash32Accumulator acc((int)TypeCategory::catAlternative);
+    int32_t hash32(instance_ptr left);
 
-        acc.add(which(left));
-        acc.add(m_subtypes[which(left)].second->hash32(eltPtr(left)));
+    instance_ptr eltPtr(instance_ptr self) const;
 
-        return acc.get();
-    }
-
-    instance_ptr eltPtr(instance_ptr self) const {
-        if (m_all_alternatives_empty) {
-            return self;
-        }
-
-        layout& record = **(layout**)self;
-
-        return record.data;
-    }
-
-    int64_t which(instance_ptr self) const {
-        if (m_all_alternatives_empty) {
-            return *(uint8_t*)self;
-        }
-
-        layout& record = **(layout**)self;
-
-        return record.which;
-    }
+    int64_t which(instance_ptr self) const;
 
     void constructor(instance_ptr self);
 
-    void destroy(instance_ptr self) {
-        if (m_all_alternatives_empty) {
-            return;
-        }
+    void destroy(instance_ptr self);
 
-        layout& record = **(layout**)self;
+    void copy_constructor(instance_ptr self, instance_ptr other);
 
-        record.refcount--;
-
-        if (record.refcount == 0) {
-            m_subtypes[record.which].second->destroy(record.data);
-            free(*(layout**)self);
-        }
-    }
-
-    void copy_constructor(instance_ptr self, instance_ptr other) {
-        if (m_all_alternatives_empty) {
-            *(uint8_t*)self = *(uint8_t*)other;
-            return;
-        }
-
-        (*(layout**)self) = (*(layout**)other);
-
-        if (*(layout**)self) {
-            (*(layout**)self)->refcount++;
-        }
-    }
-
-    void assign(instance_ptr self, instance_ptr other) {
-        layout* old = (*(layout**)self);
-
-        (*(layout**)self) = (*(layout**)other);
-
-        if (*(layout**)self) {
-            (*(layout**)self)->refcount++;
-        }
-
-        destroy((instance_ptr)&old);
-    }
+    void assign(instance_ptr self, instance_ptr other);
 
     static Alternative* Make(std::string name,
                          const std::vector<std::pair<std::string, NamedTuple*> >& types,
                          const std::map<std::string, Function*>& methods //methods preclude us from being in the memo
-                         ) {
-        if (methods.size()) {
-            return new Alternative(name, types, methods);
-        }
-
-        static std::mutex guard;
-
-        std::lock_guard<std::mutex> lock(guard);
-
-        typedef std::pair<std::string, std::vector<std::pair<std::string, NamedTuple*> > > keytype;
-
-        static std::map<keytype, Alternative*> m;
-
-        auto it = m.find(keytype(name, types));
-
-        if (it == m.end()) {
-            it = m.insert(std::make_pair(keytype(name, types), new Alternative(name, types, methods))).first;
-        }
-
-        return it->second;
-    }
+                         );
 
     const std::vector<std::pair<std::string, NamedTuple*> >& subtypes() const {
         return m_subtypes;
@@ -3041,20 +1385,7 @@ public:
         forwardTypesMayHaveChanged();
     }
 
-    bool isBinaryCompatibleWithConcrete(Type* other) {
-        if (other->getTypeCategory() == TypeCategory::catConcreteAlternative) {
-            ConcreteAlternative* otherO = (ConcreteAlternative*)other;
-
-            return otherO->m_alternative->isBinaryCompatibleWith(m_alternative) &&
-                m_which == otherO->m_which;
-        }
-
-        if (other->getTypeCategory() == TypeCategory::catAlternative) {
-            return m_alternative->isBinaryCompatibleWith(other);
-        }
-
-        return false;
-    }
+    bool isBinaryCompatibleWithConcrete(Type* other);
 
     template<class visitor_type>
     void _visitContainedTypes(const visitor_type& visitor) {
@@ -3070,12 +1401,7 @@ public:
         assert(t == m_alternative);
     }
 
-    void _forwardTypesMayHaveChanged() {
-        m_base = m_alternative;
-        m_name = m_alternative->name() + "." + m_alternative->subtypes()[m_which].first;
-        m_size = m_alternative->bytecount();
-        m_is_default_constructible = m_alternative->subtypes()[m_which].second->is_default_constructible();
-    }
+    void _forwardTypesMayHaveChanged();
 
     int32_t hash32(instance_ptr left) {
         return m_alternative->hash32(left);
@@ -3099,15 +1425,7 @@ public:
         return m_alternative->cmp(left,right);
     }
 
-    void constructor(instance_ptr self) {
-        if (m_alternative->all_alternatives_empty()) {
-            *(uint8_t*)self = m_which;
-        } else {
-            constructor(self, [&](instance_ptr i) {
-                m_alternative->subtypes()[m_which].second->constructor(i);
-            });
-        }
-    }
+    void constructor(instance_ptr self);
 
     //returns an uninitialized object of type-index 'which'
     template<class subconstructor>
@@ -3145,25 +1463,7 @@ public:
         m_alternative->assign(self, other);
     }
 
-    static ConcreteAlternative* Make(Alternative* alt, int64_t which) {
-        static std::mutex guard;
-
-        std::lock_guard<std::mutex> lock(guard);
-
-        typedef std::pair<Alternative*, int64_t> keytype;
-
-        static std::map<keytype, ConcreteAlternative*> m;
-
-        auto it = m.find(keytype(alt ,which));
-
-        if (it == m.end()) {
-            it = m.insert(
-                std::make_pair(keytype(alt,which), new ConcreteAlternative(alt,which))
-                ).first;
-        }
-
-        return it->second;
-    }
+    static ConcreteAlternative* Make(Alternative* alt, int64_t which);
 
     Type* elementType() const {
         return m_alternative->subtypes()[m_which].second;
@@ -3183,20 +1483,6 @@ private:
     int64_t m_which;
 };
 
-inline Type* Alternative::pickConcreteSubclassConcrete(instance_ptr data) {
-    uint8_t i = which(data);
-
-    return ConcreteAlternative::Make(this, i);
-}
-
-inline void Alternative::constructor(instance_ptr self) {
-    if (!m_default_construction_type) {
-        m_default_construction_type = ConcreteAlternative::Make(this, m_default_construction_ix);
-    }
-
-    m_default_construction_type->constructor(self);
-}
-
 class PythonSubclass : public Type {
 public:
     PythonSubclass(Type* base, PyTypeObject* typePtr) :
@@ -3209,19 +1495,7 @@ public:
         forwardTypesMayHaveChanged();
     }
 
-    bool isBinaryCompatibleWithConcrete(Type* other) {
-        Type* nonPyBase = m_base;
-        while (nonPyBase->getTypeCategory() == TypeCategory::catPythonSubclass) {
-            nonPyBase = nonPyBase->getBaseType();
-        }
-
-        Type* otherNonPyBase = other;
-        while (otherNonPyBase->getTypeCategory() == TypeCategory::catPythonSubclass) {
-            otherNonPyBase = otherNonPyBase->getBaseType();
-        }
-
-        return nonPyBase->isBinaryCompatibleWith(otherNonPyBase);
-    }
+    bool isBinaryCompatibleWithConcrete(Type* other);
 
     template<class visitor_type>
     void _visitContainedTypes(const visitor_type& visitor) {
@@ -3276,25 +1550,7 @@ public:
         m_base->assign(self, other);
     }
 
-    static PythonSubclass* Make(Type* base, PyTypeObject* pyType) {
-        static std::mutex guard;
-
-        std::lock_guard<std::mutex> lock(guard);
-
-        typedef std::pair<Type*, PyTypeObject*> keytype;
-
-        static std::map<keytype, PythonSubclass*> m;
-
-        auto it = m.find(keytype(base, pyType));
-
-        if (it == m.end()) {
-            it = m.insert(
-                std::make_pair(keytype(base,pyType), new PythonSubclass(base, pyType))
-                ).first;
-        }
-
-        return it->second;
-    }
+    static PythonSubclass* Make(Type* base, PyTypeObject* pyType);
 
     Type* baseType() const {
         return m_base;
@@ -3360,66 +1616,9 @@ public:
          *(PyObject**)self = buffer.getContext().deserializePythonObject(buffer);
     }
 
-    void repr(instance_ptr self, std::ostringstream& stream) {
-        PyObject* p = *(PyObject**)self;
+    void repr(instance_ptr self, std::ostringstream& stream);
 
-        PyObject* o = PyObject_Repr(p);
-
-        if (!o) {
-            stream << "<EXCEPTION>";
-            PyErr_Clear();
-            return;
-        }
-
-        if (!PyUnicode_Check(o)) {
-            stream << "<EXCEPTION>";
-            Py_DECREF(o);
-            return;
-        }
-
-        stream << PyUnicode_AsUTF8(o);
-
-        Py_DECREF(o);
-    }
-
-    char cmp(instance_ptr left, instance_ptr right) {
-        PyObject* l = *(PyObject**)left;
-        PyObject* r = *(PyObject**)right;
-
-        int res = PyObject_RichCompareBool(l, r, Py_EQ);
-        if (res == -1) {
-            PyErr_Clear();
-            if (l < r) {
-                return -1;
-            }
-            if (l > r) {
-                return 1;
-            }
-            return 0;
-        }
-
-        if (res) {
-            return 0;
-        }
-
-        res = PyObject_RichCompareBool(l, r, Py_LT);
-
-        if (res == -1) {
-            PyErr_Clear();
-            if (l < r) {
-                return -1;
-            }
-            if (l > r) {
-                return 1;
-            }
-            return 0;
-        }
-
-        if (res) {
-            return -1;
-        }
-        return 1;
-    }
+    char cmp(instance_ptr left, instance_ptr right);
 
     void constructor(instance_ptr self) {
         *(PyObject**)self = Py_None;
@@ -3441,25 +1640,7 @@ public:
         *(PyObject**)self = *(PyObject**)other;
     }
 
-    static PythonObjectOfType* Make(PyTypeObject* pyType) {
-        static std::mutex guard;
-
-        std::lock_guard<std::mutex> lock(guard);
-
-        typedef PyTypeObject* keytype;
-
-        static std::map<keytype, PythonObjectOfType*> m;
-
-        auto it = m.find(pyType);
-
-        if (it == m.end()) {
-            it = m.insert(
-                std::make_pair(pyType, new PythonObjectOfType(pyType))
-                ).first;
-        }
-
-        return it->second;
-    }
+    static PythonObjectOfType* Make(PyTypeObject* pyType);
 
     PyTypeObject* pyType() const {
         return mPyTypePtr;
@@ -3769,26 +1950,7 @@ public:
         forwardTypesMayHaveChanged();
     }
 
-    bool isBinaryCompatibleWithConcrete(Type* other) {
-        if (other->getTypeCategory() != m_typeCategory) {
-            return false;
-        }
-
-        HeldClass* otherO = (HeldClass*)other;
-
-        if (m_members.size() != otherO->m_members.size()) {
-            return false;
-        }
-
-        for (long k = 0; k < m_members.size(); k++) {
-            if (m_members[k].first != otherO->m_members[k].first ||
-                    !m_members[k].second->isBinaryCompatibleWith(otherO->m_members[k].second)) {
-                return false;
-            }
-        }
-
-        return true;
-    }
+    bool isBinaryCompatibleWithConcrete(Type* other);
 
     template<class visitor_type>
     void _visitContainedTypes(const visitor_type& visitor) {
@@ -3814,18 +1976,7 @@ public:
         }
     }
 
-    void _forwardTypesMayHaveChanged() {
-        m_is_default_constructible = true;
-        m_byte_offsets.clear();
-
-        //first m_members.size() bits (rounded to nearest byte) contains the initialization flags.
-        m_size = int((m_members.size()) / 8) + 1; //round up to nearest byte
-
-        for (auto t: m_members) {
-            m_byte_offsets.push_back(m_size);
-            m_size += t.second->bytecount();
-        }
-    }
+    void _forwardTypesMayHaveChanged();
 
     static HeldClass* Make(
             std::string inName,
@@ -3842,29 +1993,7 @@ public:
         return self + m_byte_offsets[ix];
     }
 
-    char cmp(instance_ptr left, instance_ptr right) {
-        for (long k = 0; k < m_members.size(); k++) {
-            bool leftInit = checkInitializationFlag(left,k);
-            bool rightInit = checkInitializationFlag(right,k);
-
-            if (leftInit && !rightInit) {
-                return 1;
-            }
-
-            if (rightInit && !leftInit) {
-                return -1;
-            }
-
-            if (leftInit && rightInit) {
-                char res = m_members[k].second->cmp(left + m_byte_offsets[k], right + m_byte_offsets[k]);
-                if (res != 0) {
-                    return res;
-                }
-            }
-        }
-
-        return 0;
-    }
+    char cmp(instance_ptr left, instance_ptr right);
 
     template<class buf_t>
     void deserialize(instance_ptr self, buf_t& buffer) {
@@ -3890,45 +2019,9 @@ public:
         }
     }
 
-    void repr(instance_ptr self, std::ostringstream& stream) {
-        stream << m_name << "(";
+    void repr(instance_ptr self, std::ostringstream& stream);
 
-        for (long k = 0; k < m_members.size();k++) {
-            if (k > 0) {
-                stream << ", ";
-            }
-
-            if (checkInitializationFlag(self, k)) {
-                stream << m_members[k].first << "=";
-
-                m_members[k].second->repr(eltPtr(self,k),stream);
-            } else {
-                stream << m_members[k].first << " not initialized";
-            }
-        }
-        if (m_members.size() == 1) {
-            stream << ",";
-        }
-
-        stream << ")";
-    }
-
-    int32_t hash32(instance_ptr left) {
-        Hash32Accumulator acc((int)getTypeCategory());
-
-        for (long k = 0; k < m_members.size();k++) {
-            if (checkInitializationFlag(left,k)) {
-                acc.add(1);
-                acc.add(m_members[k].second->hash32(eltPtr(left,k)));
-            } else {
-                acc.add(0);
-            }
-        }
-
-        acc.add(m_members.size());
-
-        return acc.get();
-    }
+    int32_t hash32(instance_ptr left);
 
     template<class sub_constructor>
     void constructor(instance_ptr self, const sub_constructor& initializer) const {
@@ -3945,78 +2038,22 @@ public:
         }
     }
 
-    void setAttribute(instance_ptr self, int memberIndex, instance_ptr other) const {
-        if (checkInitializationFlag(self, memberIndex)) {
-            m_members[memberIndex].second->assign(
-                eltPtr(self, memberIndex),
-                other
-                );
-        } else {
-            m_members[memberIndex].second->copy_constructor(
-                eltPtr(self, memberIndex),
-                other
-                );
-            setInitializationFlag(self, memberIndex);
-        }
-    }
+    void setAttribute(instance_ptr self, int memberIndex, instance_ptr other) const;
 
-    void emptyConstructor(instance_ptr self) {
-        //more efficient would be to just write over the bytes directly.
-        for (size_t k = 0; k < m_members.size(); k++) {
-            setInitializationFlag(self, k, false);
-        }
-    }
+    void emptyConstructor(instance_ptr self);
 
     //don't default construct classes
     static bool wantsToDefaultConstruct(Type* t) {
         return t->is_default_constructible() && t->getTypeCategory() != TypeCategory::catClass;
     }
 
-    void constructor(instance_ptr self) {
-        for (size_t k = 0; k < m_members.size(); k++) {
-            if (wantsToDefaultConstruct(m_members[k].second)) {
-                m_members[k].second->constructor(self+m_byte_offsets[k]);
-                setInitializationFlag(self, k);
-            } else {
-                setInitializationFlag(self, k, false);
-            }
-        }
-    }
+    void constructor(instance_ptr self);
 
-    void destroy(instance_ptr self) {
-        for (long k = (long)m_members.size() - 1; k >= 0; k--) {
-            if (checkInitializationFlag(self, k)) {
-                m_members[k].second->destroy(self+m_byte_offsets[k]);
-            }
-        }
-    }
+    void destroy(instance_ptr self);
 
-    void copy_constructor(instance_ptr self, instance_ptr other) {
-        for (long k = (long)m_members.size() - 1; k >= 0; k--) {
-            if (checkInitializationFlag(other, k)) {
-                m_members[k].second->copy_constructor(self + m_byte_offsets[k], other+m_byte_offsets[k]);
-                setInitializationFlag(self, k);
-            }
-        }
-    }
+    void copy_constructor(instance_ptr self, instance_ptr other);
 
-    void assign(instance_ptr self, instance_ptr other) {
-        for (long k = (long)m_members.size() - 1; k >= 0; k--) {
-            bool selfInit = checkInitializationFlag(self,k);
-            bool otherInit = checkInitializationFlag(other,k);
-
-            if (selfInit && otherInit) {
-                m_members[k].second->assign(self + m_byte_offsets[k], other+m_byte_offsets[k]);
-            }
-            else if (selfInit && !otherInit) {
-                m_members[k].second->destroy(self+m_byte_offsets[k]);
-                setInitializationFlag(self,k,false);
-            } else if (!selfInit && otherInit) {
-                m_members[k].second->copy_constructor(self + m_byte_offsets[k], other+m_byte_offsets[k]);
-                setInitializationFlag(self,k,false);
-            }
-        }
-    }
+    void assign(instance_ptr self, instance_ptr other);
 
     bool checkInitializationFlag(instance_ptr self, int memberIndex) const {
         int byte = memberIndex / 8;
@@ -4024,15 +2061,7 @@ public:
         return bool( ((uint8_t*)self)[byte] & (1 << bit) );
     }
 
-    void setInitializationFlag(instance_ptr self, int memberIndex, bool value=true) const {
-        int byte = memberIndex / 8;
-        int bit = memberIndex % 8;
-        ((uint8_t*)self)[byte] |= (1 << bit);
-
-        if (!value) {
-            ((uint8_t*)self)[byte] -= (1 << bit);
-        }
-    }
+    void setInitializationFlag(instance_ptr self, int memberIndex, bool value=true) const;
 
     const std::vector<std::pair<std::string, Type*> >& getMembers() const {
         return m_members;
@@ -4054,15 +2083,7 @@ public:
         return m_byte_offsets;
     }
 
-    int memberNamed(const char* c) const {
-        for (long k = 0; k < m_members.size(); k++) {
-            if (m_members[k].first == c) {
-                return k;
-            }
-        }
-
-        return -1;
-    }
+    int memberNamed(const char* c) const;
 
 private:
     std::vector<size_t> m_byte_offsets;
@@ -4093,15 +2114,7 @@ public:
         forwardTypesMayHaveChanged();
     }
 
-    bool isBinaryCompatibleWithConcrete(Type* other) {
-        if (other->getTypeCategory() != m_typeCategory) {
-            return false;
-        }
-
-        Class* otherO = (Class*)other;
-
-        return m_heldClass->isBinaryCompatibleWith(otherO->m_heldClass);
-    }
+    bool isBinaryCompatibleWithConcrete(Type* other);
 
     template<class visitor_type>
     void _visitContainedTypes(const visitor_type& visitor) {
@@ -4114,10 +2127,7 @@ public:
         assert(t == m_heldClass);
     }
 
-    void _forwardTypesMayHaveChanged() {
-        m_is_default_constructible = m_heldClass->is_default_constructible();
-        m_name = m_heldClass->name();
-    }
+    void _forwardTypesMayHaveChanged();
 
     static Class* Make(
             std::string inName,
@@ -4130,31 +2140,13 @@ public:
         return new Class(HeldClass::Make(inName, members, memberFunctions, staticFunctions, classMembers));
     }
 
-    instance_ptr eltPtr(instance_ptr self, int64_t ix) const {
-        layout& l = **(layout**)self;
-        return m_heldClass->eltPtr(l.data, ix);
-    }
+    instance_ptr eltPtr(instance_ptr self, int64_t ix) const;
 
-    void setAttribute(instance_ptr self, int64_t ix, instance_ptr elt) const {
-        layout& l = **(layout**)self;
-        m_heldClass->setAttribute(l.data, ix, elt);
-    }
+    void setAttribute(instance_ptr self, int64_t ix, instance_ptr elt) const;
 
-    bool checkInitializationFlag(instance_ptr self, int64_t ix) const {
-        layout& l = **(layout**)self;
-        return m_heldClass->checkInitializationFlag(l.data, ix);
-    }
+    bool checkInitializationFlag(instance_ptr self, int64_t ix) const;
 
-    char cmp(instance_ptr left, instance_ptr right) {
-        layout& l = **(layout**)left;
-        layout& r = **(layout**)right;
-
-        if ( &l == &r ) {
-            return 0;
-        }
-
-        return m_heldClass->cmp(l.data,r.data);
-    }
+    char cmp(instance_ptr left, instance_ptr right);
 
     template<class buf_t>
     void deserialize(instance_ptr self, buf_t& buffer) {
@@ -4174,15 +2166,9 @@ public:
         m_heldClass->serialize(l.data, buffer);
     }
 
-    void repr(instance_ptr self, std::ostringstream& stream) {
-        layout& l = **(layout**)self;
-        m_heldClass->repr(l.data, stream);
-    }
+    void repr(instance_ptr self, std::ostringstream& stream);
 
-    int32_t hash32(instance_ptr left) {
-        layout& l = **(layout**)left;
-        return m_heldClass->hash32(l.data);
-    }
+    int32_t hash32(instance_ptr left);
 
     template<class sub_constructor>
     void constructor(instance_ptr self, const sub_constructor& initializer) const {
@@ -4197,58 +2183,15 @@ public:
         }
     }
 
-    void emptyConstructor(instance_ptr self) {
-        if (!m_is_default_constructible) {
-            throw std::runtime_error(m_name + " is not default-constructible");
-        }
+    void emptyConstructor(instance_ptr self);
 
-        *(layout**)self = (layout*)malloc(sizeof(layout) + m_heldClass->bytecount());
+    void constructor(instance_ptr self);
 
-        layout& l = **(layout**)self;
-        l.refcount = 1;
+    void destroy(instance_ptr self);
 
-        m_heldClass->emptyConstructor(l.data);
-    }
+    void copy_constructor(instance_ptr self, instance_ptr other);
 
-    void constructor(instance_ptr self) {
-        if (!m_is_default_constructible) {
-            throw std::runtime_error(m_name + " is not default-constructible");
-        }
-
-        *(layout**)self = (layout*)malloc(sizeof(layout) + m_heldClass->bytecount());
-
-        layout& l = **(layout**)self;
-        l.refcount = 1;
-
-        m_heldClass->constructor(l.data);
-    }
-
-    void destroy(instance_ptr self) {
-        layout& l = **(layout**)self;
-        l.refcount--;
-
-        if (l.refcount == 0) {
-            m_heldClass->destroy(l.data);
-            free(*(layout**)self);
-        }
-    }
-
-    void copy_constructor(instance_ptr self, instance_ptr other) {
-        (*(layout**)self) = (*(layout**)other);
-        (*(layout**)self)->refcount++;
-    }
-
-    void assign(instance_ptr self, instance_ptr other) {
-        layout* old = (*(layout**)self);
-
-        (*(layout**)self) = (*(layout**)other);
-
-        if (*(layout**)self) {
-            (*(layout**)self)->refcount++;
-        }
-
-        destroy((instance_ptr)&old);
-    }
+    void assign(instance_ptr self, instance_ptr other);
 
     const std::vector<std::pair<std::string, Type*> >& getMembers() const {
         return m_heldClass->getMembers();
